Refactor in entities
