+++ src/main/java/com/project/loteria/config/TestConfig.java
@@ -1,18 +1,15 @@
 package com.project.loteria.config;
 
 import com.project.loteria.megasena.dtos.MSPoolDTO;
-import com.project.loteria.megasena.entities.MSBet;
+import com.project.loteria.entities.MSBet;
 import com.project.loteria.megasena.repositories.MSBetRepository;
 import com.project.loteria.megasena.service.MSBetService;
 import com.project.loteria.megasena.service.MSContestService;
 import com.project.loteria.megasena.service.MSPoolService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 @Profile("test")
