+++ src/main/java/com/project/loteria/controller/PoolController.java
@@ -1,0 +1,17 @@
+package com.project.loteria.controller;
+
+import com.project.loteria.dtos.PoolDTO;
+
+public class PoolController {
+    public void create(PoolDTO dto){
+
+    }
+
+    public void getAll(){
+
+    }
+
+    public void getAllBets(Long poolId){
+
+    }
+}
