+++ src/main/java/com/project/loteria/dtos/PoolDTO.java
@@ -1,0 +1,4 @@
+package com.project.loteria.dtos;
+
+public record PoolDTO(String title, String type) {
+}
