+++ src/main/java/com/project/loteria/entities/MSBet.java
@@ -1,4 +1,4 @@
-package com.project.loteria.megasena.entities;
+package com.project.loteria.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.project.loteria.megasena.dtos.MSBetDTO;
@@ -6,8 +6,6 @@ import com.project.loteria.service.MathService;
 import jakarta.persistence.*;
 
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 @Entity
