+++ src/main/java/com/project/loteria/megasena/repositories/MSBetRepository.java
@@ -1,6 +1,6 @@
 package com.project.loteria.megasena.repositories;
 
-import com.project.loteria.megasena.entities.MSBet;
+import com.project.loteria.entities.MSBet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MSBetRepository extends JpaRepository<MSBet, Long> {
