+++ src/main/java/com/project/loteria/megasena/repositories/MSContestRepository.java
@@ -1,6 +1,6 @@
 package com.project.loteria.megasena.repositories;
 
-import com.project.loteria.megasena.entities.MSContest;
+import com.project.loteria.entities.MSContest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MSContestRepository extends JpaRepository<MSContest, Long> {
