+++ src/main/java/com/project/loteria/megasena/repositories/MSPoolRepository.java
@@ -1,6 +1,6 @@
 package com.project.loteria.megasena.repositories;
 
-import com.project.loteria.megasena.entities.MSPool;
+import com.project.loteria.entities.MSPool;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MSPoolRepository extends JpaRepository<MSPool, Long> {
